Dynamic Configuration Updates
Enable ease of testing, and usability, by dynamically updating the
service broker configuration resource.  Modify the API to return a 503
error if configuration isn't in place or has been deleted.  Add in
synchronization so we start tests when the broker is in a known good
state.  Also allow flag parsing so we can control glog output from the
server when tests are being run.