Perform Client Cleanup
We need a way so that we can have a clean 'kubernetes' between tests to
stop one bleeding into the next.  The obvious way to do this is to use
service discovery and the dynamic client, which has highlighted a few
issues along the way with it not actually working.