Add Service Instance Tests
Basic API validation of the service instance create endpoint.  Adds in
some fixes for segfaults discovered in testing.  Fixes client resets as
the kubernetes and dynamic clients manage a mutually exclusive set of
resources.  Until we come up with a better idea caching of resources is
now banned (and caching of clients for that matter).  Perhaps shared
informers is a better solution, but then again it's low frequency so we
should just KISS.