Extended Error Checking
The OSB API specifies an error structure that is returned when things go
wrong, but is very limited in its specification of error types.  We add
in extended types that allows us to test that the error is what we'd
expect.  It also provides easy classification of whether or not the user
is to blame for screwing up e.g not us e.g. the normal case.