Improve Error Handling
There are occasions where a function may generate errors that are
classified as different (user vs configuration) so we need a nice way of
differentiating them from each other and providing canonical error
handling.  Another observation is that the HTTP error types and JSON
response are tighylu coupled, so we can drive the response completely
with an error type.