Overhaul Registry Handling
The registry has been hugely simplified and switched to using secrets.
The next step will be to make operations persist in the registry to make
the system fully stateless.  Rather than caching the whole request, it's
better to just store the pertinent things such as context and parameters
and also abstract away JSON serialization.