Remove Registry Reentrancy
To make a complex problem simpler we add a couple concepts.  A registry
item may be set to read-only, therefore cannot be committed or deleted
e.g. causing visible changes.  On top of that each go routine must use
only a single instance of a registry entry.  The entry may be cloned at
any time, but the clone will be read only.