Make Bindings Work
The registry has been changed to be based on JSON, so now we can pass
objects about.  One really good use for this is to allow a parameter to
render a credentials snippet to be returned by the API.  I've added in
password generation and updated the couchbase server example to try out
the new features.