Handle Updates Properly
Fixes #9.  Kubernetes and other things may mutate any resource we
create, to add defaults in for example, or ports for services.  We need
to be able to add and remove configuration as a result of an update and
also preserve existing stuff mutated outside of our control.  As such we
need to generate a diff of the original template, the new one and apply
it to the current resource.