Document TLS Task
TLS and networking isn't well understood by most people, so we need some
extra docuemntation around this area in order to stem the flow of
questions.  Also give some subtle hints that using CN is banned for
dynamic configuration, that's what SANs are for.  I may have to lessen
this restriction at some point if some 3rd party applications are not
compliant, but best to stake out some boundaries.