Inline Templating (#12)
While very safe defining templates in YAML, it was very verbose.  I can
foresee one of the core complaints to be this, so fix it up.  By using a
templating based approach we should get control flow (e.g. loops) for
free, and more importanlty it's not our problem, the language can be
delegated to the go documentation instead.  Implements #11.