Implicit JSON Serialization (#18)
I'm guilty of missing out the json function myself at times, so add it
in implicitly.  While this works for straight pipelines, it won't for
stuff with control flow, so outlaw this for now until we can work out
how to do this.  It will most likely involve parse trees and augmenting
leaves.