More Implicit JSON Serialization
Just taking a string on to the end of a template is very naive, it's
entrely possible some clever user will put an action in each if/else
clause and we'll miss one.  Instead walk the parse tree and augment
actions that don't set variables.  A more acceptable solution than #19
proposed.